feat issue #7
LIC-3 is implemented according to the given definition. Three unit-test methods are added to the test class for testing the functionality of the method.
Additionally, the part for double comparison in LIC-1 is changed to eliminate the double precision problem. (Since LIC-3 had the same problem test method is added for LIC-3 to test this problem)